GIF support, new commands (mahna, img2vid, deepfry), various bugfixes/adjustments

+++ video_creator.py
@@ -63,49 +63,51 @@ async def apply_filters_and_send(ctx, code, kwargs):
             vstream = input_stream.video
             astream = input_stream.audio
             vstream, astream, output_params = await code(ctx, vstream, astream, kwargs)
-            if('fs' in output_params):
-                del output_params['fs']
-            if('movflags' not in output_params):
-                output_params['movflags'] = 'faststart'
-            await set_progress_bar(ctx.message, 2)
+            if('ignore' not in output_params):
+                if('fs' in output_params):
+                    del output_params['fs']
+                if('movflags' not in output_params):
+                    output_params['movflags'] = 'faststart'
+                await set_progress_bar(ctx.message, 2)
 
-            ffmpeg_output = None
-            if(is_mp3):
-                ffmpeg_output = ffmpeg.output(astream, output_filename, **output_params)
-            elif(is_gif):
-                ffmpeg_output = ffmpeg.output(vstream, output_filename, **output_params)
-            else:
-                ffmpeg_output = ffmpeg.output(astream, vstream, output_filename, **output_params)
+                ffmpeg_output = None
+                if(is_mp3):
+                    ffmpeg_output = ffmpeg.output(astream, output_filename, **output_params)
+                elif(is_gif):
+                    ffmpeg_output = ffmpeg.output(vstream, output_filename, **output_params)
+                else:
+                    ffmpeg_output = ffmpeg.output(astream, vstream, output_filename, **output_params)
 
-            # Pass 1
-            try:
-                ffmpeg_output.run(cmd='ffmpeg-static/ffmpeg', overwrite_output=True, capture_stderr=True)
-            except ffmpeg._run.Error as e:
-                # Error will most likely happen due to the video having no audio
-                ffmpeg_output = ffmpeg.output(vstream, output_filename, **output_params)
-                ffmpeg_output.run(cmd='ffmpeg-static/ffmpeg', overwrite_output=True, capture_stderr=True)
+                # Pass 1
+                try:
+                    ffmpeg_output.run(cmd='ffmpeg-static/ffmpeg', overwrite_output=True, capture_stderr=True)
+                except ffmpeg._run.Error as e:
+                    # Error will most likely happen due to the video having no audio
+                    # TODO: add silent audio here
+                    ffmpeg_output = ffmpeg.output(vstream, output_filename, **output_params)
+                    ffmpeg_output.run(cmd='ffmpeg-static/ffmpeg', overwrite_output=True, capture_stderr=True)
 
-            # Pass 2 (if the file is too big)
-            resulting_filesize = os.path.getsize(output_filename) / 1000000
-            if(resulting_filesize > 7.9):
-                # Calculate bitrate needed
-                longest_duration = 0
-                metadata = FFProbe(output_filename)
-                for stream in metadata.streams:
-                    if(stream.is_video() or stream.is_audio()):
-                        duration = stream.duration_seconds()
-                        longest_duration = max(longest_duration, duration)
-                output_params['b:v'] = 7500000 / longest_duration
+                # Pass 2 (if the file is too big)
+                resulting_filesize = os.path.getsize(output_filename) / 1000000
+                if(resulting_filesize > 7.9):
+                    # Calculate bitrate needed
+                    longest_duration = 0
+                    metadata = FFProbe(output_filename)
+                    for stream in metadata.streams:
+                        if(stream.is_video() or stream.is_audio()):
+                            duration = stream.duration_seconds()
+                            longest_duration = max(longest_duration, duration)
+                    output_params['b:v'] = 7500000 / longest_duration
 
-                # Create the new video
-                input_filename_pass2 = 'vids/pass2' + output_filename.split('/')[1]
-                os.rename(output_filename, input_filename_pass2)
-                input_stream = ffmpeg.input(input_filename_pass2)
-                ffmpeg.output(input_stream, output_filename, **output_params).run(cmd='ffmpeg-static/ffmpeg', overwrite_output=True, capture_stderr=True)
-                os.remove(input_filename_pass2)
+                    # Create the new video
+                    input_filename_pass2 = 'vids/pass2' + output_filename.split('/')[1]
+                    os.rename(output_filename, input_filename_pass2)
+                    input_stream = ffmpeg.input(input_filename_pass2)
+                    ffmpeg.output(input_stream, output_filename, **output_params).run(cmd='ffmpeg-static/ffmpeg', overwrite_output=True, capture_stderr=True)
+                    os.remove(input_filename_pass2)
 
 
-            await set_progress_bar(ctx.message, 3)
+                await set_progress_bar(ctx.message, 3)
             try:
                 await ctx.send(file=discord.File(output_filename))
             except Exception as e:
@@ -141,7 +143,7 @@ async def apply_corruption_and_send(ctx, code, code_kwargs, avi_kwargs = {}, mp4
     await set_progress_bar(ctx.message, 0)
     input_vid, is_yt, result = await media_cache.download_last_video(ctx)
     if(not result):
-        await ctx.send("Error downloading the video")
+        await ctx.send("There was an error downloading the video, try uploading the video again.")
         return
     await set_progress_bar(ctx.message, 1)
